Use two buffer rings in buffer mode
In a multi-window configuration, switching between a source buffer and
the REPL is a common workflow. This formerly would cause the (global)
buffer ring to reflect these two buffers as being the most recent,
which would mean that traversing the ring in the source buffer window
would cause it to visit the REPL on its way to whatever other source
buffer was most recent before that.

We now use two rings to avoid navigating to irrelevant buffers here,
where the two rings correspond to the two most common workflow
categories of content / source buffers and messaging / I/O buffers.