Simplify setting face font attribute
An error came up where `face-attribute` returned a font-object instead
of a spec, which then threw an error when it was passed to `font-get`;
however, it seems that `set-face-font` can take a string, font spec,
font entity, or font object.

This change utilizes `set-face-font`'s versatility by first trying to
find a non-nil result of `find-font` (a font entity), then using
`face-attribute` to get the default font (a font object, or 'unspecified).