Basic working mode rings
Rings are a natural structure to use for modes. This has been planned
for a long time and this is a basic working version that we can
iterate on. It replaces modes with mode rings at a particular level in
a specific tower (lisp tower, so that we can rotate normal → symex →
normal), and stays behind black box interfaces.