Use Lithium instead of Hydra in View mode
Hydra was useful to prototype Rigpa modes, but we need more lifecycle
hooks in order to ensure clean mode entry and exit. There were a
number of hacks in Chimera (the modal interface abstraction layer that
was introduced specifically to abstract such hacks and present a
simple modal API) to achieve an approximation of the expected behavior
with mode transitions, but there were still some corner cases that
could not be addressed.

With Lithium, we use Emacs's built-in minor modes to implement the
more rigid Vim-style (and yet, "point-free," Symex-style) modes we're
looking for. This allows us to benefit from all of the existing
infrastructure available for minor modes, and keeps the overlying
additions "lightweight." By having an in-house, dedicated, modal
interface framework, we also gain the flexibility to add the lifecycle
hooks we need and tailor the implementation, without having to work
around design choices that may have been made for requirements
different from our own.