Rename `enter-mode` → `switch-mode` to highlight transition
This runs exit hooks for the previous mode and entry hooks for the new
mode. We also don't run any hooks for Lithium modes which do all their
hooks internally and via their exposed API. We currently use
`manage-hooks` as a proxy for "is this a Lithium mode?" Just a
temporary bridge to something more explicit.