Incorporate wrapping hooks into low-level mode transitions
We recently fixed an infinite recursion where post-exit hooks enter
the "appropriate" mode which calls the main "switch mode" interface,
which itself attempts to exit the current mode before entering the
next, and thus triggers post-exit hooks again. To address that, we had
the "appropriate" mode entry function call the low-level entry
function instead of the main switch mode transition function, but this
meant that these were no longer triggering the pre-entry hooks as part
of mode entry.

This moves the pre-entry and post-exit hooks into the low-level entry
and exit functions, respectively, so that they are correctly triggered
in either of these cases (while still avoiding the original infinite
loop since we only _enter_ and do not _exit_ (again) as part of
post-exit hooks, while still doing both in the high level "switch").

We also need to explicitly call low level mode exit in on-demand mode
transitions like "enter lower" and "escape higher" when in a non-local
starting mode.