On exiting foreign modes, only set the local evil state
We no longer rely on the evil state as the source of truth on what the
current mode is (except for evil-backed states, which currently are
just the built-in evil states). But we do still rely on evil for UI
feedback, to indicate visually which mode we are in, even for modes
that are not evil-backed. We do this just because there is existing
infrastructure for this for Evil, in the form of mode line
enhancements and perhaps other things. Eventually we will likely want
to move away from this, but for now, the only thing we still need to
do upon exiting forreign modes is to enter the appropriate evil state
for UI feedback purposes, which we do now instead of attempting to
enter any particular local mode (which we don't need to do since there
would still be a local mode on the stack upon foreign mode exit).