escape to normal from replace state
While in replace state, escaping was having no effect. As it is
considered a foreign state, it attempts to only exit the mode, which
for evil-backed modes is a no-op. Additionally, as it isn't a lithium
mode, there isn't even a notion of a stack wherein exiting replace
mode would reinstate any prior mode. Escaping to normal seems the most
straightforward thing to do, but I'm not sure if there may be a more
correct handling.