(history) Fix setting evil state upon entry and exit
This still doesn't show History mode in the actual read-only
timemachine buffer on initial entry (but does, on subsequent re-entry
into the mode), and it might be that that buffer is being created
after the entry hooks have been run. But that is a relatively
superficial issue, and this does fix the unclean exit in the source
buffer.

Note that the evil state is only used for visual feedback, to allow
leveraging existing Evil ecosystem tools for the purpose.