move creation of editing structures to user config, for now
This allows us to evaluate this config after both rigpa and
symex-rigpa have loaded. A better approach would be for Rigpa to
provide a "core language" of modes (not including symex) and a handful
of core editing structures for the ground level and meta levels, and
for modes like symex to be added to this core set as user extensions,
via either third party packages or in user config. We likely want the
editing structures to be dynamic variables defined in Rigpa, and which
are modified by these extensions.